[BDA][NFC]: use SSA construction algorithm.
This new implementation reduces the disjoint path problem to SSA construction.
The old disjoint path implementation was based on Ford-Fulkerson and a
reduction of the disjoint path problem to MAXFLOW.
The new implementation is a simple data flow analysis and takes linear time (#blocks) per
call to join_blocks. The old implementation had quadratic complexity.